Factorized tests as well.
